+++ app/components/training_with_disability_component.html.erb
