+++ app/views/candidate_interface/training_with_a_disability/_form.html.erb
@@ -21,7 +21,7 @@
 <%= f.govuk_radio_buttons_fieldset(
   :disclose_disability,
   legend: { text: t('application_form.training_with_a_disability.disclose_disability.label'), size: 'm' },
-  hint: { text: t('application_form.training_with_a_disability.disclose_disability.hint') }
+  hint: { text: t('application_form.training_with_a_disability.disclose_disability.hint') },
 ) do %>
   <%= f.govuk_radio_button :disclose_disability, 'yes', label: { text: t('application_form.training_with_a_disability.disclose_disability.yes') }, link_errors: true do %>
     <%= f.govuk_text_area :disability_disclosure, rows: 8, label: { text: t('application_form.training_with_a_disability.disability_disclosure.label'), size: 's' }, max_words: 400 %>
