+++ app/views/support_interface/application_forms/show.html.erb
@@ -47,7 +47,7 @@
 
 <%= render PersonalStatementComponent.new(application_form: @application_form, editable: @application_form.editable?) %>
 
-<%= render ProviderInterface::TrainingWithDisabilityComponent.new(application_form: @application_form) %>
+<%= render TrainingWithDisabilityComponent.new(application_form: @application_form) %>
 
 <%= render InterviewPreferencesComponent.new(application_form: @application_form) %>
