+++ spec/components/training_with_disability_component_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe ProviderInterface::TrainingWithDisabilityComponent do
+RSpec.describe TrainingWithDisabilityComponent do
   context 'when the candidate has not disclose disability support' do
     it 'renders that no help is required' do
       application_form = instance_double(
@@ -9,7 +9,7 @@ RSpec.describe ProviderInterface::TrainingWithDisabilityComponent do
         disability_disclosure: nil,
       )
       result = render_inline(described_class.new(application_form:))
-      expect(result.text).to include('Do you want to ask for help to become a teacher?No')
+      expect(result.text).to include('Do you want to tell providers about a support need?No')
       expect(result.text).not_to include('Give any relevant information')
     end
   end
@@ -22,8 +22,8 @@ RSpec.describe ProviderInterface::TrainingWithDisabilityComponent do
         disability_disclosure: 'I am hard of hearing',
       )
       result = render_inline(described_class.new(application_form:))
-      expect(result.text).to include('Do you want to ask for help to become a teacher?Yes, I want to share information about myself so my provider can take steps to support me')
-      expect(result.text).to include('Give any relevant informationI am hard of hearing')
+      expect(result.text).to include('Do you want to tell providers about a support need?Yes')
+      expect(result.text).to include('Details of your support needsI am hard of hearing')
     end
   end
 
@@ -35,7 +35,7 @@ RSpec.describe ProviderInterface::TrainingWithDisabilityComponent do
         disability_disclosure: '',
       )
       result = render_inline(described_class.new(application_form:))
-      expect(result.text).to include('Do you want to ask for help to become a teacher?No')
+      expect(result.text).to include('Do you want to tell providers about a support need?No')
       expect(result.text).not_to include('Give any relevant information')
     end
   end
